requirements.txt
# =============================================================================
# NANO BANANA STUDIO PRO - PYTHON DEPENDENCIES v2.0
# =============================================================================
# pip install -r requirements.txt
# =============================================================================

# =============================================================================
# CORE FRAMEWORK
# =============================================================================
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# =============================================================================
# ASYNC & NETWORKING
# =============================================================================
httpx>=0.26.0
aiofiles>=23.2.1
aiohttp>=3.9.0
websockets>=12.0
sse-starlette>=1.8.0

# =============================================================================
# DATABASE & CACHING
# =============================================================================
redis>=5.0.0
hiredis>=2.3.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
alembic>=1.13.0

# =============================================================================
# MEDIA PROCESSING - VIDEO
# =============================================================================
ffmpeg-python>=0.2.0
imageio>=2.33.0
imageio-ffmpeg>=0.4.9
moviepy>=1.0.3
av>=11.0.0

# =============================================================================
# MEDIA PROCESSING - IMAGE
# =============================================================================
Pillow>=10.2.0
opencv-python>=4.9.0.80
opencv-python-headless>=4.9.0.80
scikit-image>=0.22.0

# =============================================================================
# FACE DETECTION & RECOGNITION
# =============================================================================
mediapipe>=0.10.9
insightface>=0.7.3
onnxruntime>=1.16.0
# onnxruntime-gpu>=1.16.0  # Uncomment for GPU support

# =============================================================================
# AUDIO PROCESSING
# =============================================================================
librosa>=0.10.1
soundfile>=0.12.1
pydub>=0.25.1
aubio>=0.4.9
scipy>=1.12.0
numpy>=1.26.0

# =============================================================================
# AI / ML - CORE
# =============================================================================
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
transformers>=4.36.0
accelerate>=0.25.0
safetensors>=0.4.0
huggingface-hub>=0.20.0

# =============================================================================
# AI / ML - DIFFUSERS (Image/Video Generation)
# =============================================================================
diffusers>=0.25.0
xformers>=0.0.23  # Optional: Memory-efficient attention
bitsandbytes>=0.42.0  # Optional: 8-bit quantization

# =============================================================================
# AI / ML - AUDIO GENERATION
# =============================================================================
audiocraft>=1.3.0  # MusicGen
bark>=0.1.5  # Suno Bark TTS
TTS>=0.22.0  # Coqui TTS / XTTS
SunoAI>=1.0.0  # Suno API client (pip install SunoAI)

# =============================================================================
# AI / ML - SUNO INTEGRATION
# =============================================================================
tenacity>=8.2.0  # Retry logic for API calls
# 2captcha-python>=1.2.0  # Optional: CAPTCHA solving

# =============================================================================
# AI / ML - SPEECH RECOGNITION
# =============================================================================
openai-whisper>=20231117
faster-whisper>=0.10.0  # Faster Whisper inference

# =============================================================================
# LLM INTEGRATION
# =============================================================================
openai>=1.10.0
anthropic>=0.18.0
google-generativeai>=0.3.0
tiktoken>=0.5.0

# =============================================================================
# YOUTUBE PUBLISHING
# =============================================================================
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.118.0

# =============================================================================
# DOCUMENT PROCESSING
# =============================================================================
PyMuPDF>=1.23.0  # PDF processing
python-docx>=1.1.0  # Word documents
python-pptx>=0.6.23  # PowerPoint
markdown>=3.5.0
pyyaml>=6.0.1
jinja2>=3.1.3

# =============================================================================
# DATA & UTILITIES
# =============================================================================
pandas>=2.1.0
numpy>=1.26.0
tqdm>=4.66.0
tenacity>=8.2.0
orjson>=3.9.0
jsonschema>=4.21.0

# =============================================================================
# LOGGING & MONITORING
# =============================================================================
loguru>=0.7.2
structlog>=24.1.0
prometheus-client>=0.19.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0

# =============================================================================
# TESTING
# =============================================================================
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0  # For testing FastAPI

# =============================================================================
# DEVELOPMENT
# =============================================================================
black>=24.1.0
ruff>=0.1.14
mypy>=1.8.0
isort>=5.13.0
pre-commit>=3.6.0

# =============================================================================
# OPTIONAL - COMFYUI INTEGRATION
# =============================================================================
# These are installed separately in ComfyUI environment
# comfy-cli>=1.0.0
# custom-nodes for ComfyUI

# =============================================================================
# OPTIONAL - GPU ACCELERATION
# =============================================================================
# Uncomment based on your GPU:
# For NVIDIA CUDA:
# torch>=2.1.0+cu121
# torchvision>=0.16.0+cu121
# torchaudio>=2.1.0+cu121
# onnxruntime-gpu>=1.16.0

# For AMD ROCm:
# torch>=2.1.0+rocm5.6
# torchvision>=0.16.0+rocm5.6
# torchaudio>=2.1.0+rocm5.6
